refactor: restructure Shared module to views directory
  - Move Shared module from src/app/share/views/SharedLinksView to src/views/Shared
  - Organize components into dedicated subdirectories (components, hooks, context, utils)
  - Move ShareItemDialog and EmptySharedView to Shared module components
  - Update all import paths to reference new module structure
  - Improve code organization with proper separation of concerns
  - Add accessibility improvements to PasswordInput (use native button element)
  - Fix deprecated onKeyPress usage in TextInput
  - Add proper error handling in sharedViewUtils
