+++ src/views/Shared/components/EmptySharedView.tsx
@@ -1,8 +1,8 @@
 import { UploadSimple, Users } from '@phosphor-icons/react';
-import { useTranslationContext } from '../../../i18n/provider/TranslationProvider';
-import Empty from '../../../shared/components/Empty/Empty';
-import folderEmptyImage from 'assets/icons/light/folder-open.svg';
-import { SharedNamePath } from 'app/share/types';
+import { useTranslationContext } from '../../../app/i18n/provider/TranslationProvider';
+import Empty from '../../../app/shared/components/Empty/Empty';
+import folderEmptyImage from '../../../assets/icons/light/folder-open.svg';
+import { SharedNamePath } from '../../../app/share/types';
 
 interface EmptySharedViewProps {
   onUploadFileButtonClicked: () => void;
