+++ src/views/Shared/context/SharedViewContext.ts
@@ -1,7 +1,7 @@
 import { createContext, Dispatch } from 'react';
-import { OrderDirection } from '../../../../core/types';
-import { DriveItemData } from '../../../../drive/types';
-import { AdvancedSharedItem, PreviewFileItem } from '../../../types';
+import { OrderDirection } from '../../../app/core/types';
+import { DriveItemData } from '../../../app/drive/types';
+import { AdvancedSharedItem, PreviewFileItem } from '../../../app/share/types';
 import { OrderField } from '../components/SharedItemList';
 
 export interface ShareViewState {
