+++ src/views/Shared/index.tsx
@@ -1,0 +1,6 @@
+export { default as SharedViewWrapper } from './SharedViewWrapper';
+export { default as SharedView } from './SharedView';
+export * from './components/SharedItemList';
+export * from './components/ShareItemDialog/ShareItemDialog';
+export * from './hooks/useFetchSharedData';
+export * from './hooks/useSharedContextMenu';
